[Platinum V] Title: 등산 마니아, Time: 636 ms, Memory: 105504 KB, Score: 42 point -BaekjoonHub

+++ 백준/Platinum/20188. 등산 마니아/등산 마니아.java
@@ -41,13 +41,13 @@ public class Main {
 
 //        visited[1] = true;
         dfs(1);
-        int result = 0;
+        long result = 0;
         int pair = (n * (n-1)) / 2;   //7 6 5 4 3 2 1   8C2 = (8 * 7) / 2  28 * 7
         for(int i = 2; i <= n; i++){
             int existPair = n - subtree[i];
             int calc = pair - ((existPair * (existPair-1)) / 2);
 //            System.out.println(calc + " ");
-            result += calc;
+            result += (long)calc;
         }
         System.out.println(result);
     }
